TODAY I CHANGED THE FOLLOWING: 1. Updated com.revature.models.reimbursement by uncommenting my timestamp fields ReimbursementAccepted and ReimbursementResolve. Imported org.postgresql.jdbc.TimestampUtils and the corresponding types, but asked batchmates and they tod me that it was java.sql.Timestamp, so I changed them to that instead. 2. Fixed the setUserRole by adding appropriate parameters (String UserRole). 3. Refactored Reimb_Status, Reimb_Type, and User_Role to ReimbStatus, ReimbType, and UserRole in keeping with Java naming conventions. 4. Changed final line of IUserDAO and UserDAO from "public boolean addUserWithHome" to "public boolean addUserWithReimbursement".
TO-DO NEXT:
a. clean up the UserController Class. It still has all of the DB column names from the homes table (from the Avengers demo) and change those blocks of code to the columns from the reimbursement table.
b. Add those same columns to both the Users and Reimbursement tables (and the other three tables).
c. Add Dummy Data to all tables to ensure things are working correctly.
d. Work on the frontend of this project and make sure that basic functionality and communication is complete for frontend. Work on backend and middleware over the weekend.
e. If by some miracle, all of that is done, then work on building test methods for the service layer.